Large scale cleanup of warning level #4 messages.  This primarily involved 3 areas: 1.  Variables that were initialized, but never used 2.  Variables of one type (such as int) being assigned to a smaller type (such as byte).  These were explicitly cast to duplicate what the compiler has been doing all along.  This also included signed types being assigned to unsigned types. 3.  Unreachable code.  There were 2 or 3 places in the code where an explicit and unconditional return statement was coded, but was immediately followed by more code.  These sections were commented out.
There are still a few places with level 4 warnings that still need to be improved, but some of these will take a little study.
Issue/Tracker #3076157


git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@355 41062d70-b26d-460a-9302-f438a8930db8

+++ src/DrawPin.cpp
@@ -263,8 +263,8 @@ CDrawPin::CDrawPin(CTinyCadDoc *pDesign)
   m_elec=0;
   m_str="";
   m_show=3;
-  m_length = pDesign->GetOptions()->GetPinLength();
-  m_part = m_pDesign->IsEditLibrary() ? static_cast<CTinyCadSymbolDoc*>(m_pDesign)->GetPart() : 0;
+  m_length = (WORD) pDesign->GetOptions()->GetPinLength();
+  m_part = (BYTE) (m_pDesign->IsEditLibrary() ? static_cast<CTinyCadSymbolDoc*>(m_pDesign)->GetPart() : 0);
   m_number_pos = pDesign->GetOptions()->GetPinNumberPos();
   m_centre_name = FALSE;
 }
@@ -807,7 +807,7 @@ void CDrawPin::Rotate(CDPoint p,int ndir)
   m_point_a = CDPoint(m_point_a.x+p.x,m_point_a.y+p.y);
   m_point_b = CDPoint(m_point_b.x+p.x,m_point_b.y+p.y);
 
-  m_dir = DoRotate(m_dir,ndir);
+  m_dir = (BYTE) DoRotate(m_dir,ndir);
 }
 
 const TCHAR* CDrawPin::GetElectricalTypeName(int i)
